Make json runtype impure
After the isPure check was added to some of the test helpers, json runtype tests were failing. This happened because the json runtype was set to always be pure, but it is actually always impure -- it doesn't return the string value originally passed in, but returns the JSON object version of the string.

The json runtype is now always impure.