Account for primitive value equality in tests
For both undefinedOr and nullOr, their 'deals with impure runtypes' test was using `expectAcceptValuesPure`, which isn't correct because the runtypes under test were not pure. `expectAcceptValuesImpure` couldn't be used as a replacement because it assumes that the outputted value after applying the runtype will not identical to the original value passed to the runtype. However, the outputted value will be identical if no modification (e.g. trim) took place and the value is a primitive value (e.g. undefined, 'foo', 1, true).

These tests have been updated to not use either helper, but to check runtype output directly.